Fixing bug in source.elliptical
When generating a elliptical galaxy with a shift, part of the galaxy is outside of the generated image, leading to a smaller object area and wrong normalization. Net effect is a higher surface brightness of the source.

Solution: Generated galaxy image is now centered on the image and shifting is done when generating the source.

source.spiral also fixed.